Decouple config file from HashStore, move it to 'FileHashStore' module and rename to 'filehashstore_config'
